Add `cargo-saw-build` and `saw-rustc` commands
Because `cargo-saw-build` shares so much code in common with `cargo-crux-test`,
I have factored out the common code into a new `cargo_test_common.rs` file.

Fixes #39.
